bigquery: support sending envetkit data directly to bigquery
Current architecture enables saving evenkit events to BigQuery (BQ) via eventkitd-bigquery daemon:

 1. eventkit sends a UDP message to a server
 2. eventkitd-bigquery server collects UDP messages and sends them to BQ

There are use-cases (like evetkit-time), where only a few messages should be saved to BQ, but reliability is prefered.

In this case, it can be more reliable to save data directly to BQ, especialy if the environment is trusted and the Google credentialse can be shared.

This patch:

 * separates the logic which updates scheme / saves data to BQ from the eventkit-daemon logic
 * provides an addiional destination implementation (direct save)
 * adds a new helper to test the new functionality
