+++ eventkitd-bigquery/go.mod
@@ -20,8 +20,8 @@ require (
 	github.com/google/uuid v1.3.0 // indirect
 	github.com/googleapis/enterprise-certificate-proxy v0.1.0 // indirect
 	github.com/googleapis/gax-go/v2 v2.5.1 // indirect
-	github.com/jtolio/eventkit/eventkitd v0.0.0-20221107155359-4ea4ead7a706 // indirect
 	github.com/klauspost/compress v1.15.11 // indirect
+	github.com/pkg/errors v0.9.1 // indirect
 	go.opencensus.io v0.23.0 // indirect
 	go.uber.org/atomic v1.10.0 // indirect
 	go.uber.org/multierr v1.8.0 // indirect
