+++ tools/go.work
@@ -3,3 +3,4 @@ go 1.18
 use ..
 use .
 use ../eventkitd
+use ../eventkitd-bigquery
